Comments and Exceptions included
-In Lan I ask Yewo a question
-In the other two I basically leave out the usual comments asking
questions and cleaning up the code a little bit
-The rest of the work was putting scalfolding for exception handling. We
are currently experiencing a bug that is described in TCPClient in
RefreshAction(). I manage to throw the exceptions where I think there
might be problems and catch them in the refresh action where I think the
problem will need to be resolved at.
-After this point I plan on making a development branch and continue
working from there