Exception handling and comments on Reconnect
-So I did the usual comments on Reconnecting page.
-But I also included exception handling in the case that it fails to
connect to the server. By extension I did the same for if it is
connecting for the first time.
-Most importantly, I think I fixed a bug by making a method thats
recursively called get out of the loop and start again. There was a
return keyword missing within TCPClient when it calls CreateClient in
itself. That might have been causing a deadlock

-@Yewo you are welcome to test my work at this stage but I feel like I
have to finish somemore handling that I left over from the other day.