Added Exception handling for FailedtoRetrieveData
-The handleing was interesting to do cause essentially I figured there
might be a situation when the awaitresponse was set by the dataRecieved
method but was forced into null by the 	SendRequest Method when it is
first listening for awaitresponse. So I added if statements to make sure
it checks if it is still not changed by using the processingRequest bool
and is it is still the same as the old value then it will set
awaitresponse to null. We should consider just using old and new values
in a refactor.
-I also removed a comment inside ForcedToLogOutPopupPage